pyproject.toml
[project]
name = "telecom-anomaly-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.2",
    "fpdf2>=2.8.7",
    "jpype1>=1.6.0",
    "matplotlib>=3.4.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.0.0",
    "seaborn>=0.11.0",
    "tabula-py>=2.10.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "black>=21.0.0",
    "flake8>=3.9.0",
    "pytest>=6.2.0",
]
